docs(Dropdown): improve code samples
