+++ packages/core/src/components/Dropdown/stories/Empty.tsx
@@ -1,0 +1,3 @@
+import { HvDropdown } from "@hitachivantara/uikit-react-core";
+
+export const Empty = () => <HvDropdown aria-label="Empty" />;
